requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
browserforge==1.2.3
build==1.3.0
cachetools==6.2.2
camoufox==0.4.11
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.3.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
courlan==1.3.2
cssselect==1.3.0
curl-cffi==0.13.0
dateparser==1.2.2
distro==1.9.0
django==5.2.7
durationpy==0.10
eval-type-backport==0.2.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
htmldate==1.9.4
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.1.4
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
invoke==2.2.1
jieba3k==0.35.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
justext==3.0.2
kubernetes==34.1.0
langdetect==1.0.9
language-tags==1.2.0
lxml==6.0.2
lxml-html-clean==0.4.3
markdown-it-py==4.0.0
mdurl==0.1.2
mistralai==1.9.11
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.19.0
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pillow==12.0.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
playwright==1.55.0
posthog==5.4.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyee==13.0.0
pygments==2.19.2
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
readability-lxml==0.8.4.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-file==3.0.0
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
scrapling==0.3.7
screeninfo==0.8.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.22.1
tqdm==4.67.1
trafilatura==2.0.0
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
